# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Acestream Scraper API",
    "description": "API for scraping and managing Acestream channels",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/playlists/m3u": {
      "get": {
        "tags": ["playlists"],
        "summary": "Get M3U Playlist",
        "description": "Generate M3U playlist with specified filters\n\n- **search**: Optional search term for channel names\n- **group**: Optional specific group to filter by\n- **only_online**: Whether to include only online channels (default: True)\n- **include_groups**: Comma-separated list of groups to include\n- **exclude_groups**: Comma-separated list of groups to exclude",
        "operationId": "get_m3u_playlist_api_v1_playlists_m3u_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Search"
            }
          },
          {
            "name": "group",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Group"
            }
          },
          {
            "name": "only_online",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Only Online"
            }
          },
          {
            "name": "include_groups",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Include Groups"
            }
          },
          {
            "name": "exclude_groups",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Exclude Groups"
            }
          },
          {
            "name": "base_url",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Base Url"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "text/plain": { "schema": { "type": "string" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/warp/mode": {
      "post": {
        "tags": ["warp"],
        "summary": "Set WARP mode",
        "description": "Set the WARP mode (warp, dot, proxy, or off).\n\nArgs:\n    request: The mode to set WARP to.\n\nReturns:\n    Success message if the mode was set successfully.",
        "operationId": "set_warp_mode_api_v1_warp_mode_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/WarpModeRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/warp/license": {
      "post": {
        "tags": ["warp"],
        "summary": "Register WARP license",
        "description": "Register a license key with the WARP client.\n\nArgs:\n    request: The license key to register.\n\nReturns:\n    Success message if the license was registered successfully.",
        "operationId": "register_warp_license_api_v1_warp_license_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
      "WarpModeEnum": {
        "type": "string",
        "enum": ["warp", "dot", "proxy", "off"],
        "title": "WarpModeEnum",
        "description": "Available WARP modes for API use"
      },
      "WarpModeRequest": {
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/WarpModeEnum",
            "description": "The WARP mode to set"
          }
        },
        "type": "object",
        "required": ["mode"],
        "title": "WarpModeRequest",
        "description": "Request schema for changing WARP mode"
      }
    }
  }
}